% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_dimensions.R
\name{add_pos_0_on_Leaflet}
\alias{add_pos_0_on_Leaflet}
\title{Add position 0 on leaflet}
\usage{
add_pos_0_on_Leaflet(x)
}
\arguments{
\item{x}{The Area data.frame}
}
\value{
The Area data.frame with the 0 position for each leaflet if missing, so Area
will potentially have more rows
}
\description{
Add a position 0 to the leaflet positions
}
\examples{
\dontrun{
add_pos_0_on_Leaflet(Area)
}
}
