% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AlKashi}
\alias{AlKashi}
\title{AlKashi}
\usage{
AlKashi(x, l)
}
\arguments{
\item{x}{lenght of one side}

\item{l}{lenght of the second side}
}
\value{
The length of the third side
}
\description{
Get the length of the 3 side of a triangle knowing the 2 other sides using
the law of cosines, also known as the generalized Pythagorean theorem.
}
