% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_archi.R
\name{compute_archi}
\alias{compute_archi}
\title{Compute Architectural parameters}
\usage{
compute_archi(map, data_path, write_path = NULL)
}
\arguments{
\item{map}{Palm trees age desired in months after planting}

\item{data_path}{Path to the folder were all data files resides}

\item{write_path}{Path were to write the function outputs. If \code{NULL} (the default),
the data is not written.}
}
\value{
Returns invisibly the input data and the models fit as a list for each progeny.
}
\description{
Imports the data and fit the models to compute the architectural
parameters for VPalm.
}
\note{
Uses \code{\link[=import_data]{import_data()}} and \code{\link[=mod_all]{mod_all()}} under the hood
}
