% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vars.R
\name{compute_MAP}
\alias{compute_MAP}
\title{Compute MAP for development data}
\usage{
compute_MAP(x)
}
\arguments{
\item{x}{The development data.frame}
}
\value{
The development data.frame with new or updated MAP.
}
\description{
Compute the months after planting for the development data from
the transplanting date and the observation date. The function also corrects the
MAP if the filed campaign is made over 2 consecutive months (should have the same MAP).
}
\examples{
\dontrun{
compute_MAP(development)
}
}
