% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{curvature}
\alias{curvature}
\title{Curvature}
\usage{
curvature(position, coefCurv)
}
\arguments{
\item{position}{The positions along the axis relative to the tip (A point)}

\item{coefCurv}{Curvature coefficient}
}
\value{
The curvature angles for each position
}
\description{
Compute the angles of points from relative positions and curvature
}
\examples{
\dontrun{
library(Vpalmr)
curvature(position= 0.2, coefCurv= 0.5)
}
}
