% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_curvature.R
\name{curvature_error}
\alias{curvature_error}
\title{Leaf curvature error}
\usage{
curvature_error(angC, position, Length, X_pos, Y_pos, coefCurv, decAInfl)
}
\arguments{
\item{angC}{The angle of the leaf at C point}

\item{position}{The position of each point along the rachis relative to the tip (A point)}

\item{Length}{The total rachis length}

\item{X_pos}{The observed X positions in centimeters}

\item{Y_pos}{The observed Y positions in centimeters}

\item{coefCurv}{Curvature coefficient}

\item{decAInfl}{The inflection point of the sigmoidal curve that predict point A}
}
\value{
The error of prediction of the leaf curvature
}
\description{
Compute the leaf curvature as VPalm do, compare
the output coordinates to observations, and return the error. This function is
generally used for parameter optimization
}
\details{
The error is computed as the sum of the summed squarred root errors
of X and Y coordinates
}
\note{
The function must be applied to each leaf independently
}
\examples{
\dontrun{
library(Vpalmr)
curvature_error(angC,position,Length,coefCurv,decAInfl)
}
}
