% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ops.R
\name{design_plot}
\alias{design_plot}
\title{Design the plot}
\usage{
design_plot(nRow, nCol, x0, y_distance, x_distance)
}
\arguments{
\item{nRow}{Number of tree rows}

\item{nCol}{Number of tree columns}

\item{x0}{The minimum X coordinates}

\item{y_distance}{The intra-row distance}

\item{x_distance}{The inter-row distance}
}
\value{
A data.frame with all information to make an Open Plant Scene file.
}
\description{
Help designing the planting pattern of the plot
}
\details{
For the moment the function only creates triangular planting pattern,
but more options will be added with time
}
\examples{
design_plot(nRow = 4, nCol = 5, x0 = 0, y_distance = 9.2,
            x_distance = 8)

}
