% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_curvature.R
\name{deviation_A}
\alias{deviation_A}
\title{Leaf deviation}
\usage{
deviation_A(devC, Length)
}
\arguments{
\item{devC}{Deviation on each point compared to C point}

\item{Length}{The total rachis length}
}
\value{
The leaf deviation in A point
}
\description{
Compute the leaf deviation on point A
}
\examples{
\dontrun{
library(Vpalmr)
deviation_A(devC= 86, Lenght= 320)
}
}
