% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{distance_2D}
\alias{distance_2D}
\title{Distance bewteen 2D points}
\usage{
distance_2D(x, y)
}
\arguments{
\item{x}{The values in the x axis}

\item{y}{The values in the y axis}
}
\value{
The total distance (lenght) of the series of objects
}
\description{
Estimate the linear distance between points in 2D coordinates
}
\details{
The function returns a 0 value for the first two points.
}
\examples{
\dontrun{
library(Vpalmr)
distance_2D(x= 1:10,y= 1:10)
}
}
