% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_archi.R
\name{estimate_archi}
\alias{estimate_archi}
\title{Estimate architectural parameters}
\usage{
estimate_archi(progeny, map, Dec)
}
\arguments{
\item{progeny}{Character. The progeny name}

\item{map}{Integer. The desired age of the Palm for simulation, in Month After Planting.}

\item{Dec}{Declination angle data.frame for all progenies}

\item{output}{The path to the folder where to write the outputs}
}
\value{
VPalm inputs
}
\description{
Estimate palm architectural parameters from field measurements, and export
them as VPalm inputs.
}
\details{
This function is highly adapted to a given input format. Please make sure to make
the input formats as required.
}
\examples{
\dontrun{
library(Vpalmr)
estimate_archi(progeny= "DY",map= 47)
}
}
