% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraxt_raw.R
\name{extract_form}
\alias{extract_form}
\title{Extract form}
\usage{
extract_form(form)
}
\arguments{
\item{form}{The form (usually a sheet from an excel file)}
}
\value{
A data.frame with the different measurements
}
\description{
Extract the data from a form of one leaf from one Tree
}
\examples{
extract_form(form)
}
