% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw.R
\name{extract_form_dev}
\alias{extract_form_dev}
\title{Extract development form}
\usage{
extract_form_dev(form)
}
\arguments{
\item{form}{The data form on which extraction is made}
}
\value{
A data.frame with the data from all sheets (all leaves)
}
\description{
Extract the data from a leaf development form. This function is called
by \code{\link[=extract_sheets]{extract_sheets()}}. Users should use \code{\link[=extract_sheets]{extract_sheets()}} instead of this function.
}
\examples{
\dontrun{
extract_form_dev(form)
}
}
