% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_list.R
\name{extract_list}
\alias{extract_list}
\title{VPalm parameter list for all progenies}
\usage{
extract_list(data, model, nb_leaves = 45, seed = sample.int(1000, 1),
  init = init_list())
}
\arguments{
\item{data}{A list of all data (generally from \code{\link[=import_data]{import_data()}})}

\item{model}{A list of the models fitted on the data (generally from \code{\link[=mod_all]{mod_all()}})}

\item{nb_leaves}{The number of leaves needed}

\item{seed}{The seed for random parameter generation}

\item{init}{Initialisation values (see \code{\link[=extract_params]{extract_params()}})}
}
\value{
A list of several VPalm parameters lists, one for each progeny, and one
with the average (non-NA) values.
}
\description{
Make a VPalm parameter list for each progeny provided in the
data from the \link[=mod_stem_diameter]{models} outputs using \code{\link[=extract_params]{extract_params()}} and replace
missing value by the average of all progenies.
}
