% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_list.R
\name{extract_progeny}
\alias{extract_progeny}
\title{VPalm parameter list for all progenies}
\usage{
extract_progeny(
  data,
  model,
  n,
  leaves = 45,
  seed = NULL,
  init = init_list(),
  average = T
)
}
\arguments{
\item{data}{A list of all data (generally from \code{\link[=import_data]{import_data()}})}

\item{model}{A list of the models fitted on the data (generally from \code{\link[=mod_all]{mod_all()}})}

\item{n}{The number of samplings for each progeny, or in other words, the
number of trees to be generated for each progeny}

\item{leaves}{The number of leaves needed}

\item{seed}{The seeds for random parameter generation, see details.}

\item{init}{Initialisation values (see \code{\link[=extract_trees]{extract_trees()}})}

\item{average}{Boolean, is mean of all average trees to be used to replace missing values ?}
}
\value{
A nested list, containing a list of VPalm parameters for each tree sampled
for each Progeny, and the average tree of each progeny, and one with the average (non-NA) values for all
progenies. Each list in the list has as many n lists that represents each sampled tree
for each progeny.
}
\description{
Make a VPalm parameter list for n trees sampled in the parameters
distributions of each progeny, and also for an average tree from the n trees distribution.
provided in the data from the \link[=mod_stem_diameter]{models}
outputs using \code{\link[=extract_params]{extract_params()}} and replace missing value by the average of all progenies.
}
\details{
\code{seed} should be a named list of length equal to the number of progenies, and each
object with n seeds each. The names of the list objects should match the ones of the
progenies. Alternatively, \code{seed} can be a vector of seeds of length n that will be recycled
for each progeny. If \code{NULL}, a random seed is generated for each tree of each progeny
using \code{\link[base:sample]{base::sample()}}.
}
