% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraxt_raw.R
\name{extract_section}
\alias{extract_section}
\title{Extract section}
\usage{
extract_section(df_archi, Row, Col)
}
\arguments{
\item{Row}{The row index of the section}

\item{Col}{The column index of the section}
}
\value{
A data.frame with the section form data
}
\description{
Extract a section from the form
}
\details{
The row and column of the section is the one from the upper left corner
}
\examples{
extract_section(5,1)
}
