% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraxt_raw.R
\name{extract_sheets}
\alias{extract_sheets}
\title{Extract sheets}
\usage{
extract_sheets(path, sheet = NULL)
}
\arguments{
\item{path}{The path to the excel file}

\item{sheet}{The sheets index on the files (see details)}
}
\value{
A data.frame with the data from all sheets (all leaves)
}
\description{
Extract the data from all forms and all sheets from the leaf area and
biomass excel files
}
\details{
If the sheet is not given, the function will import data from all sheets. Otherwise,
the user can provide a given sheet, or a vector of sheets (see examples).
}
\examples{
# Extract all sheets:
extract_sheets(path= "1-Data/raw/Form Archi.xlsx")

# Extract only the first sheet:
extract_sheets(path= "1-Data/raw/Form Archi.xlsx", sheet= 1)

# Extract the second and third sheets:
extract_sheets(path= "1-Data/raw/Form Archi.xlsx", sheet= c(2,3))

}
