% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_params.R
\name{extract_trees}
\alias{extract_trees}
\title{Extract several VPalm parameters}
\usage{
extract_trees(
  data,
  model,
  n,
  leaves = 45,
  seed = sample.int(1000, n),
  init = init_list()
)
}
\arguments{
\item{data}{A list of all data (generally from \code{\link[=import_data]{import_data()}})}

\item{model}{A list of the models fitted on the data (generally from
\code{\link[=mod_all]{mod_all()}})}

\item{n}{The number of trees desired (how many samples are made)}

\item{leaves}{The number of leaves required}

\item{seed}{The seed for random parameter generation (must be length n)}

\item{init}{The initialisation parameters (see details)}
}
\value{
n+1 list of the parameters used as VPalm input, one for each n tree, and
one for the average tree
}
\description{
Extract Vpalm parameters for n trees and for the average tree using
\code{\link[=extract_params]{extract_params()}} recursively
}
\details{
n controls the number of trees sampled in the parameter value
distribution. Parameters sampled include:
\itemize{
\item The rachis curvature
\item The rachis relative height
\item All lme models (see \link[=mod_stem_diameter]{models})
\item All nlme models (see \link[=mod_stem_diameter]{models})
}

The init parameter should be a named list of initialization
parameters as in \code{\link[=init_list]{init_list()}}. The user can use \code{\link[=init_list]{init_list()}} as a
template to change the values of the init argument (see examples).

Missing values are replaced by values from the average tree
}
