% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ops.R
\name{format_ops}
\alias{format_ops}
\title{Format ops information}
\usage{
format_ops(design, Progeny, map, id = 1, bounds = TRUE,
  pavement = NULL, average = FALSE)
}
\arguments{
\item{design}{The planting design, generally computed using \code{\link[=design_plot]{design_plot()}}}

\item{Progeny}{The progeny name}

\item{map}{The tree age in month after planting}

\item{id}{The scene ID}

\item{bounds}{Are the plot bounds written in the OPS file ? Used for backward
compatibility.}

\item{pavement}{Pavement file path. Only used if the pavement file (\code{.gwa}) has to
be linked in the OPS. Used for backward compatibility.}

\item{average}{Boolean. Use average tree instead of the sampled ones.}
}
\value{
A pre-formatted OPS
}
\description{
Make the planting design to the Open Plant Scene format
to prepare for OPS writing
}
