% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_param}
\alias{format_param}
\title{Format a VPalm input list}
\usage{
format_param(data)
}
\arguments{
\item{data}{An \code{\link[=extract_params]{extract_params()}} list, or one from the \code{\link[=extract_list]{extract_list()}} lists}
}
\value{
A string in the VPalm input format
}
\description{
Format an \code{\link[=extract_params]{extract_params()}} list or one of the \code{\link[=extract_list]{extract_list()}} list
to prepare for writting
}
\examples{
\dontrun{
VPalm_out= Vpalmr::extract_list(data= Inputs, model= models, nb_leaves= 45, seed= 10)
average_params= format_param(data = VPalm_out$Average)
}
}
