% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_progeny}
\alias{format_progeny}
\title{Format all parameters lists}
\usage{
format_progeny(data)
}
\arguments{
\item{data}{A \code{\link[=extract_progeny]{extract_progeny()}} output}
}
\value{
A list of VPalm inputs
}
\description{
Format all the lists from \code{\link[=extract_progeny]{extract_progeny()}} output to prepare for writting.
}
\details{
This function applies \code{\link[=format_tree]{format_tree()}} sequentially to all trees from all
progenies and to the average list.
}
\examples{
\dontrun{
VPalm_out= Vpalmr::extract_list(data= Inputs, model= models, nb_leaves= 45, seed= 10)
out= format_list(data = VPalm_out)
}

}
