% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_output.R
\name{format_tree}
\alias{format_tree}
\title{Format a VPalm input list}
\usage{
format_tree(data)
}
\arguments{
\item{data}{An \code{\link[=extract_params]{extract_params()}} list, or one from the \code{\link[=extract_progeny]{extract_progeny()}} lists}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with two columns:
\itemize{
\item name: the Java-formatted variable initialization for VPalm
\item value: the variable value
}
}
\description{
Format an \code{\link[=extract_params]{extract_params()}} list or one of the \code{\link[=extract_progeny]{extract_progeny()}} list
to prepare for writting
}
\examples{
\dontrun{
VPalm_out= Vpalmr::extract_progeny(data= Inputs, model= models, nb_leaves= 45, seed= 10)
average_params= format_param(data = VPalm_out$Average)
}
}
