% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_VPalm.R
\name{format_VPalm}
\alias{format_VPalm}
\title{Format VPalm inputs}
\usage{
format_VPalm(data)
}
\arguments{
\item{data}{A \code{\link[=VPalm_list]{VPalm_list()}} output}
}
\value{
A string in the VPalm input format
}
\description{
Format the \code{\link[=VPalm_list]{VPalm_list()}} output to be written in a VPalm
input file
}
