% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Sigmoidal function}
\usage{
import_data(file_archi, file_la, file_phylo, file_angles)
}
\arguments{
\item{file_archi}{Path to the architectural file}

\item{file_la}{Path to the leaf area file}

\item{file_phylo}{Path to the phylotaxy file}

\item{file_angles}{Path to the leaf declination angles file}
}
\value{
A sigmoid
}
\description{
Estimate a sigmoidal function from parameters.
}
