% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{leaf_curvature}
\alias{leaf_curvature}
\title{Leaf curvature}
\usage{
leaf_curvature(position, angC, angA, coefCurv, Length)
}
\arguments{
\item{position}{The positions along the rachis relative to the tip (A point)}

\item{angC}{The angle of the leaf at C point}

\item{angA}{The angle of the leaf at A point (tip)}

\item{coefCurv}{Curvature coefficient}

\item{Length}{The total rachis length}
}
\value{
The leaf curvature as a data.frame of the X and Y coordinates of the input relative positions
}
\description{
Compute the leaf curvature as VPalm do
}
\examples{
\dontrun{
library(Vpalmr)
leaf_curvature(position= c(0,0.43,0.69,1),
               angC= c(27.14,27.14,27.14,27.14),
               angA= c(66.59,66.59,66.59,66.59),
               coefCurv= 3.58,
               Lenght= c(320.11,320.11,320.11,320.11))
}
}
