% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{leaflet_axial_angle}
\alias{leaflet_axial_angle}
\title{Axial insertion angle}
\usage{
leaflet_axial_angle(position, angle_C, slope_C, angle_A)
}
\arguments{
\item{position}{Relative position of the leaflet}

\item{angle_C}{Angle of the leaflet compared to C point}

\item{slope_C}{Slope of the leaflet to C point relationship}

\item{angle_A}{Angle of the leaflet compared to A point}
}
\value{
The leaflet axial angle
}
\description{
Computes the leaflet axial insertion angle
}
