% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_dimensions.R
\name{leaflet_length}
\alias{leaflet_length}
\alias{leaflet_max_width}
\alias{leaflet_max_width}
\title{Leaflet dimensions}
\usage{
leaflet_length(X, Ymax, Y0, Yfin, X_Ymax)

leaflet_max_width(X, Ymax, Y0, Yfin, X_Ymax)
}
\arguments{
\item{X}{Relative position along the leaf}

\item{Ymax}{Maximal Y}

\item{Y0}{Intercept}

\item{Yfin}{Value at the maximal X}

\item{X_Ymax}{X value at Ymax}
}
\value{
The leaflet length or maximum width
}
\description{
Computes the leaflet length or maximum width using leaflet
position along the leaf and several parameters
}
