% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_disposition.R
\name{leaflet_position}
\alias{leaflet_position}
\title{Leaflet position}
\usage{
leaflet_position(Rank, param)
}
\arguments{
\item{Rank}{Leaflet rank on rachis}

\item{param}{Distribution parameter}
}
\value{
The leaflet position on the rachis
}
\description{
Predict the leaflet position along the rachis
using the leaflet relative rank on rachis and a parameter
}
