% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{leaflet_radial_angle}
\alias{leaflet_radial_angle}
\title{Radial insertion angle}
\usage{
leaflet_radial_angle(position, A0, Amax, Xm)
}
\arguments{
\item{position}{Relative position of the leaflet}

\item{A0}{Intercept}

\item{Amax}{Maximum angle}

\item{Xm}{Maximum position value}
}
\value{
The leaflet axial angle
}
\description{
Computes the leaflet radial insertion angle
}
