% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_dimensions.R
\name{leaflet_shape_nls}
\alias{leaflet_shape_nls}
\title{Fit nls on leaflet shape}
\usage{
leaflet_shape_nls(width, position)
}
\arguments{
\item{width}{Leaflet relative width (compared to maximum observed on the leaf, 0-1)}

\item{position}{Leaflet relative position on the leaf (0-1)}
}
\value{
The estimation of the xm and ym parameters from the nls fit
}
\description{
Find the coefficients of the beta function for leaflet shape.
}
\details{
The function determines the nonlinear (weighted) least-squares estimates of the
xm and ym parameters using the \code{\link[stats]{nls}} of the following model:
\code{width~shape_beta(x= position,xm= xm,ym= ym)}.
See \code{\link{shape_beta}} for more informations.
}
