% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_list.R
\name{make_list}
\alias{make_list}
\title{VPalm parameter list for all progenies}
\usage{
make_list(data, model, nb_leaves = 45, seed = sample.int(1000, 1))
}
\arguments{
\item{data}{A list of all data (generally from \code{\link[=import_data]{import_data()}})}

\item{model}{A list of the models fitted on the data (generally from \code{\link[=mod_all]{mod_all()}})}

\item{nb_leaves}{The number of leaves needed}

\item{seed}{The seed for random parameter generation}
}
\value{
A list of two:
\itemize{
\item Progeny, a list of VPalm parameters for each progeny
\item Average, A list of VPalm parameters with average parameter values
}
}
\description{
Make a VPalm parameter list for each progeny provided in the
data from the \link[=mod_stem_diameter]{models} outputs using \code{\link[=VPalm_list]{VPalm_list()}} and replace
missing value by the average of all progenies.
}
