% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_opf.R
\name{make_opf}
\alias{make_opf}
\title{Make OPF}
\usage{
make_opf(parameter, opf, AMAPStudio, overwrite = T)
}
\arguments{
\item{parameter}{The VPalm parameter file path and name (see \code{\link[=write_progeny]{write_progeny()}}
or \code{\link[=write_tree]{write_tree()}})}

\item{opf}{The target OPF file path and name}

\item{AMAPStudio}{The root path to AMAPStudio}

\item{overwrite}{Boolean. Should pre-existing OPF files overwriten ?}
}
\value{
Writes an OPF file
}
\description{
Use VPalm from AMAPStudio to make an Open Plant Format file using
a parameter file generally created by \code{\link[=write_progeny]{write_progeny()}} or \code{\link[=write_tree]{write_tree()}}.
}
