% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ops.R
\name{make_ops_all}
\alias{make_ops_all}
\title{Format and write all OPS}
\usage{
make_ops_all(Progeny, design, map, path)
}
\arguments{
\item{path}{}
}
\value{
An OPS for each progeny, each design and each map.
}
\description{
Format and write OPS from a set of design experiment(s),
map and progenies, by applying \code{\link[=format_ops]{format_ops()}} and \code{\link[=write_ops]{write_ops()}} sequentially
}
\details{
The function uses \code{\link[base:mapply]{base::mapply()}} to apply both \code{\link[=format_ops]{format_ops()}} and
\code{\link[=write_ops]{write_ops()}} to any number of progeny, design, or map. So if these arguments
are provided with equal length, they will be applied in parallel (\emph{i.e.} in a
multivariate mode) as \code{\link[base:mapply]{base::mapply()}} do. Each can also have length one, so it
will be re-cycled for each combination.
}
