% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mod_stem_diameter}
\alias{mod_stem_diameter}
\alias{mod_stem_height}
\alias{mod_rachis_length}
\alias{mod_petiole_ratio}
\alias{mod_B_position}
\alias{mod_nb_leaflet}
\alias{mod_C_declination}
\alias{mod_leaf_curvature}
\alias{mod_A_declination}
\title{Fit the models}
\usage{
mod_stem_diameter(data)

mod_stem_height(data)

mod_rachis_length(data)

mod_petiole_ratio(data, Physio_age)

mod_B_position(data, Physio_age)

mod_nb_leaflet(data)

mod_C_declination(data)

mod_leaf_curvature(data)

mod_A_declination(data, decMaxA = 180, decSlopeA = 0.01)
}
\arguments{
\item{data}{Input data, generally one of the objects from \code{\link[=import_data]{import_data()}}}

\item{decMaxA}{The maximum allowed declination angle at the A point}

\item{decSlopeA}{The declination slope at point A}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of the model outputs, eventually grouped by Progeny
if several in input

@importFrom dplyr group_by do mutate filter
}
\description{
These functions fit a given model from which the user can
extract parameters used as VPalm inputs
}
\section{Purpose of each function}{

\itemize{
\item \code{\link[=mod_stem_diameter]{mod_stem_diameter()}}: model the stem diameter from the rachis length using a
\code{\link[=sigmoid]{sigmoid()}}
\item \code{\link[=mod_stem_height]{mod_stem_height()}}: model the stem height using the total emitted
leaves throughout the palm life using \code{\link[=stem_height]{stem_height()}}
\item \code{\link[=mod_rachis_length]{mod_rachis_length()}}: model the leaf length from the leaf number (index of the
leaves starting from the first leaf ever emitted)
\item \code{\link[=mod_petiole_ratio]{mod_petiole_ratio()}}: model the petiole/rachis ratio by taking the average value for
the progeny at the given physiological age +/- 60. If there is no value, return the
average value by progeny for all leaves older than the reference leaf rank (>17).
\item \code{\link[=mod_nb_leaflet]{mod_nb_leaflet()}}: model the number of leaflets from the rachis length using a
\code{\link[=sigmoid]{sigmoid()}}
\item \code{\link[=mod_C_declination]{mod_C_declination()}}: model the leaf declination angle at C point from leaf rank
\item \code{\link[=mod_leaf_curvature]{mod_leaf_curvature()}}: find the leaf inclination parameters (coefCurv and decAInfl)
by minimizing the difference between a computed leaf using these parameters and the
measurements. Uses \code{\link[=optim_Leaf_Curv_Infl]{optim_Leaf_Curv_Infl()}} under the hood.
\item \code{\link[=mod_A_declination]{mod_A_declination()}}: model the leaf declination angle at A point from C angle. This
function needs the decAInfl parameter computed by \code{\link[=mod_leaf_curvature]{mod_leaf_curvature()}}, so it first
checks if the parameter is present in the input data, and computes it if absent. The user
should avoid this step by providing the parameter when possible because the
\code{\link[=mod_leaf_curvature]{mod_leaf_curvature()}} call can be computationally intensive.
}
}

\section{The models}{

\itemize{
\item \code{\link[stats:nls]{stats::nls()}} is used for \code{\link[=mod_stem_diameter]{mod_stem_diameter()}}, \code{\link[=mod_stem_height]{mod_stem_height()}}, \code{\link[=mod_nb_leaflet]{mod_nb_leaflet()}}
\item \code{\link[nlme:lme]{nlme::lme()}} is used for \code{\link[=mod_rachis_length]{mod_rachis_length()}}, \code{\link[=mod_C_declination]{mod_C_declination()}}
\item \code{\link[nlme:nlme]{nlme::nlme()}} is used for
\item \code{\link[base:mean]{base::mean()}} is used for \code{\link[=mod_petiole_ratio]{mod_petiole_ratio()}}
\item \code{\link[stats:optim]{stats::optim()}} is used by \code{\link[=mod_leaf_curvature]{mod_leaf_curvature()}}
}
}

