% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{object_tans}
\alias{object_tans}
\title{Compute the tangants along non-linear objects}
\usage{
object_tans(x, y, res = 1, method = c("loess", "smooth.spline"), ...)
}
\arguments{
\item{x}{The values in the x axis}

\item{y}{The values in the y axis}

\item{res}{The re-sampling resolution, with the same unit as inputs.}

\item{method}{The smoothing method, either using \code{\link[stats:loess]{stats::loess()}} or \code{\link[stats:smooth.spline]{stats::smooth.spline()}}}

\item{...}{Further arguments to pass to the smoothing function}
}
\value{
The tangent for each point of a curved object
}
\description{
Estimate the tangants for points from a curved object using smoothing
}
\details{
The x and y coordinates must be ordered from first to last point on the object.
The function is used to compute the tangents of points for a curved object. The function
re-sample the object using a smoothing function to give the closest tangant to the point.
The higher the resolution, the higher the sub-sampling, and the closer the tangent will be.
Here is how the function works:
\enumerate{
\item The user sample some points from a curved object and register their X and Y positions
in a euclidean coordinate system, and give the points coordinates as input of the function
\item An approximation of the object curvature is made using a smoothing function on all
measured points
\item The object is re-sampled using the resulting smoothing function with a
sampling resolution of \code{res}.
\item The tangent is computed for each input point using the closer re-sampled points
}
}
\examples{
\dontrun{
library(Vpalmr)
object_tans(x= 1:10,y= rep(1,10))
}
}
\seealso{
\code{\link[=object_lenght]{object_lenght()}}
}
