% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{optim_Leaf_Curv_Infl}
\alias{optim_Leaf_Curv_Infl}
\title{Optimize leaf curvature parameters}
\usage{
optim_Leaf_Curv_Infl(data)
}
\arguments{
\item{data}{The input data.frame}
}
\value{
A data.frame with coefCurv and decAInfl values, the final error value returned by
\code{\link[=curvature_error]{curvature_error()}} and the convergence value from \code{\link[stats:optim]{stats::optim()}}.
}
\description{
Optimize coefCurv and decAInfl, two leaf curvature parameters from VPalm
using field data
}
\details{
The input data.frame must have the following columns:
\itemize{
\item Rank: the leaf rank. If not available, at least an index or a grouping factor for each leaf.
\item angC: the leaf C angle
\item RelativePositionRachisEstim: the relative position of the given points in the leaf compared
to the tip.
\item rachisLength: total length of the leaf
\item X_distance_cm: X coordinates in centimeters
\item Y_distance_cm: Y coordinates in centimeters
}
}
\examples{
\dontrun{
library(Vpalmr)
optim_Leaf_Curv_Infl(data)
}
}
