% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_all}
\alias{plot_all}
\title{Plot all model outputs}
\usage{
plot_all(data, model, nb_leaves = 45)
}
\arguments{
\item{data}{A list of all data (generally from \code{\link[=import_data]{import_data()}})}

\item{model}{The models fitted on the data (generally from any \link[=mod_stem_diameter]{model})}

\item{nb_leaves}{The number of leaves to plot}
}
\value{
A list of all plots
}
\description{
Plot all model outputs
}
