% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_params}
\alias{plot_params}
\title{Plot the architectural parameters estimations}
\usage{
plot_params(progeny, pdfdir)
}
\arguments{
\item{progeny}{Character. The progeny name}

\item{pdfdir}{The path to the folder where to write the plots}
}
\value{
A series of plots
}
\description{
Plot the estimations of the palm architectural parameters from field measurements
for control
}
\examples{
\dontrun{
library(Vpalmr)
archi_param= estimate_archi(progeny= "DY",map= 47)
plot_params(archi_param)
}
}
