% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_stem_diameter}
\alias{plot_stem_diameter}
\alias{plot_stem_height}
\alias{plot_rachis_length}
\alias{plot_petiole_ratio}
\alias{plot_B_position}
\alias{plot_nb_leaflet}
\alias{plot_C_declination}
\alias{plot_leaf_curvature}
\alias{plot_A_declination}
\title{Plot the architectural parameters estimations}
\usage{
plot_stem_diameter(data, model)

plot_stem_height(data, model)

plot_rachis_length(data, model)

plot_petiole_ratio(data, model, Physio_age)

plot_B_position(data, model, Physio_age)

plot_nb_leaflet(data, model)

plot_C_declination(data, model)

plot_leaf_curvature(data, model)

plot_A_declination(data, model)
}
\arguments{
\item{data}{Input data, generally one of the objects from \code{\link[=import_data]{import_data()}}}

\item{model}{Model fitted from one of \link[=mod_stem_diameter]{models}}

\item{Physio_age}{Physiological age, in number of leaves emmitted from birth}
}
\value{
A ggplot
}
\description{
Plot the estimations of the palm architectural parameters from
their parameter estimation and input data.
}
\section{Plotting functions (plot_*)}{

\itemize{
\item \code{\link[=plot_stem_diameter]{plot_stem_diameter()}}: plot the predicted stem diameter at the level of
the considered leaf using the leaf length
\item \code{\link[=plot_stem_height]{plot_stem_height()}}: plot the predicted stem height using the total cumulated
emitted leaves
\item \code{\link[=plot_rachis_length]{plot_rachis_length()}}: plot the predicted leaf length according to its index (leaf
number)
\item \code{\link[=plot_petiole_ratio]{plot_petiole_ratio()}}: plot the observed petiole to rachis ratio along the average
value used for simulation.
\item \code{\link[=plot_nb_leaflet]{plot_nb_leaflet()}}: plot the predicted number of leaflets in function of rachis
length
\item \code{\link[=plot_C_declination]{plot_C_declination()}}: plot the declination angle at C point in function of leaf rank
\item \code{\link[=plot_leaf_curvature]{plot_leaf_curvature()}}: plot the leaf curvature in function of the relative position
on the leaf
\item \code{\link[=plot_A_declination]{plot_A_declination()}}; plot the leaf declination angle at A point in function of the C point
}

@importFrom dplyr group_by do mutate filter
@importFrom tidyr unnest
@importFrom ggplot2 ggplot facet_wrap geom_point geom_line ylab xlab aes
}

\examples{
\dontrun{
library(Vpalmr)
archi_param= estimate_archi(progeny= "DY",map= 47)
plot_params(archi_param)
}
}
