% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pmean}
\alias{pmean}
\title{Parallel mean}
\usage{
pmean(..., na.rm = T)
}
\arguments{
\item{...}{Numeric arguments}

\item{na.rm}{A logical indicating whether missing values should be removed.}
}
\value{
A vector of the parallel mean of same length as the arguments.
}
\description{
Returns the parallel mean, in the same fashion as \code{\link[base:Extremes]{base::Extremes()}}
}
\details{
The function uses a combination of \code{\link[base:cbind]{base::cbind()}} and \link[base:colSums]{rowMeans}.
The function can take one or more vectors or matrices as arguments and returns a single
'parallel' mean of the vectors. The first element of the result is the mean of the first
elements of all the arguments and so on.
}
\examples{
pmean(1:10, 21:30)
}
