% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{position_on_leaflet}
\alias{position_on_leaflet}
\title{Positions on leaflet}
\usage{
position_on_leaflet(x)
}
\arguments{
\item{x}{The (maximum) length of the leaflet}
}
\value{
A vector of 8 positions along the leaflet
}
\description{
Compute positions along the leaflet length
}
\examples{
position_on_leaflet(25)
}
