% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pull_lme}
\alias{pull_lme}
\title{Pull lme output}
\usage{
pull_lme(data, epsilon = 10^-6)
}
\arguments{
\item{data}{The output from one of the \link[=mod_stem_diameter]{lme models}}

\item{epsilon}{epsilon value for matrix inversion}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with model outputs
}
\description{
Format lme outputs in a standard \code{\link[tibble:tibble]{tibble::tibble()}} and computes
some usefull informations such as covariance matrices. Carefull, this function is
poorly designed for the moment and is only applicable on the lme used in this context
because only two factors can be used.
}
\details{
Epsilon is used to avoid non semi-positive matrices, and is added to the matrix
diagonal to make it closer to the identity matrix
}
