% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf.R
\name{section_height}
\alias{section_height}
\title{Rachis section height}
\usage{
section_height(Position, a)
}
\arguments{
\item{Position}{Relative position on the rachis (0-1)}

\item{a}{Parameter}
}
\value{
The section height
}
\description{
Computes the leaf section height from its
relative position on the leaf and the \code{a} parameter.
}
