% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_dimensions.R
\name{shape_beta}
\alias{shape_beta}
\title{Beta function for leaflet shape}
\usage{
shape_beta(x, xm, ym)
}
\arguments{
\item{x}{Relative position on the leaf}

\item{xm}{Position of the maximum leaflet width (0-1)}

\item{ym}{Shape factor}
}
\value{
The shape of the leaflet
}
\description{
The function computes the shape of the leaflet according
to its position on the leaf and two parameters: xm and ym.
}
