% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoidal function}
\usage{
sigmoid(X, max, slope, infl)
}
\arguments{
\item{X}{The x values}

\item{max}{The y maximum value}

\item{slope}{The slope at the inflection point}

\item{infl}{The X position of the inflection point}
}
\value{
A sigmoid
}
\description{
Estimate a sigmoidal function from parameters.
}
\examples{
\dontrun{
library(Vpalmr)
sigmoid(X= 1/10,max= 3,slope= 1,infl=5)
}
}
