% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem.R
\name{Stem_height}
\alias{Stem_height}
\title{Stem height}
\usage{
Stem_height(X, y0, coef)
}
\arguments{
\item{X}{Total number of leaves emmitted throughout the palm life}

\item{y0}{Minimum height}

\item{coef}{Growth coefficient}
}
\value{
The stem height
}
\description{
Computes the stem height from the total number of leaves emitted during
the palm life cycle
}
\details{
The stem height is computing following this equation:
\deqn{Stemheight=y_0\cdot e^{coef\cdot X}}{Stem_height= y0*exp(coef*X)}
}
