% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inputs.R
\name{test_Area}
\alias{test_Area}
\alias{test_development}
\title{Test the input data}
\usage{
test_Area(x, path = NULL)

test_development(x, path = NULL)
}
\arguments{
\item{x}{The input data.frame}

\item{path}{The path to a directory were to write the plots}
}
\value{
\code{NULL} if no major issue was found in the file, or a character string of all
cumulated potential errors in the data.frame.
}
\description{
Performs several tests on the VPalm input files to check for potential
errors such as wrong format, missing essential data or typos.
}
\details{
The outputs of the functions are used to make global warnings, messages or errors
}
\examples{
\dontrun{
library(Vpalmr)
Area= data.table::fread('LeafArea_monitoring_SMSE.csv',data.table = F,dec=',')
test_Area(Area= Area)
}
}
