% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inputs.R
\name{test_pos_on_leaflet}
\alias{test_pos_on_leaflet}
\title{Test the position on leaflet}
\usage{
test_pos_on_leaflet(data)
}
\arguments{
\item{data}{A sample of the Area data file}
}
\value{
The input data.frame with one more column for the test: \code{err}. If the section
pass the test, \code{err} have the value "no error", else it will take a character value of
the given error.
}
\description{
Test the position on leaflet (\code{PositionOnLeaflet}) format and
values from one section of a leaf in a given palm tree.
}
\examples{
\dontrun{
# Data frame with the right positions:
df_sub= data.frame(TreeNumber= "90_15", LeafIndex= -1, Section= 1,
                   PositionOnLeaflet= Vpalmr::position_on_leaflet(56))
# Testing (all should be ok):
test_pos_on_leaflet(df_sub)

# Adding a wrong value:
df_sub[5,'PositionOnLeaflet']= 20
test_pos_on_leaflet(df_sub)
}
}
