% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{up_progress}
\alias{up_progress}
\title{Update progress bar}
\usage{
up_progress(prog, name)
}
\arguments{
\item{prog}{A progress function}

\item{name}{The progress info (e.g. name of the function executed)}
}
\value{
Nothing. Just update the prog function.
}
\description{
Update progress bar
}
\details{
This function is a helper that is called to increment a progress function
passed as an argument. It is used for example for a Shiny progress bar using
\href{https://shiny.rstudio.com/articles/progress.html}{progress}.
}
