% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{warn_inc}
\alias{warn_inc}
\title{warning construction}
\usage{
warn_inc(warn, x)
}
\arguments{
\item{x}{A previous warning message as a character string}
}
\value{
One warning from two: the warn object and x
}
\description{
Helper to add warnings and return only one
}
\keyword{internal}
