% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_VPalm.R
\name{write_list}
\alias{write_list}
\title{Write VPalm parameter lists}
\usage{
write_list(data, path = getwd(), name = NULL)
}
\arguments{
\item{data}{A \code{\link[=format_list]{format_list()}} output}

\item{path}{The path of the directory to write the files to.}

\item{name}{(Optionnal) The names of the files. If \code{NULL}, the names of the progenies
will be used}
}
\description{
Writes all the lists from \code{\link[=format_list]{format_list()}} as VPalm input files.
}
\details{
This function applies \code{\link[=write_params]{write_params()}} sequentially to all progenies and to
the average list.
}
\examples{
\dontrun{
out= format_list(data = VPalm_out)
write_list(data = out)
}

}
