% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_VPalm.R
\name{write_models}
\alias{write_models}
\title{Write model}
\usage{
write_models(data, path, name = "models")
}
\arguments{
\item{data}{Model outputs, generally from \code{\link[=mod_all]{mod_all()}}}

\item{path}{Path to the folder where to write it}

\item{name}{Name of the file to write}
}
\description{
Writes the models outputs to disk.
}
\details{
The file can be read back using \code{\link[base:readRDS]{base::readRDS()}}
}
