% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ops.R
\name{write_ops}
\alias{write_ops}
\title{Write OPS}
\usage{
write_ops(data, path, overwrite = T)
}
\arguments{
\item{data}{A pre-formated OPS file, generally the output from \code{\link[=format_ops]{format_ops()}}}

\item{path}{File path and name}

\item{overwrite}{Boolean. Should pre-existing OPS files overwriten ?}
}
\value{
Writes an OPS to disk, and returns \code{TRUE} if successfull or \code{FALSE} otherwise.
}
\description{
Write OPS
}
