% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_VPalm.R
\name{write_param}
\alias{write_param}
\title{Write VPalm parameter}
\usage{
write_param(data, path = getwd())
}
\arguments{
\item{data}{A \code{\link[=format_param]{format_param()}} output (or one of the \code{\link[=format_list]{format_list()}})}

\item{path}{The path of the VPalm input file}
}
\description{
Writes a parameter list from \code{\link[=format_param]{format_param()}} (or one from
\code{\link[=format_list]{format_list()}}) as a VPalm input files.
}
\examples{
\dontrun{
out= format_list(data = VPalm_out)
write_param(data = out$Average)
}

}
