% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_VPalm.R
\name{write_tree}
\alias{write_tree}
\title{Write VPalm tree parameter file}
\usage{
write_tree(data, path = getwd(), name = "vpalm_input", age = NULL)
}
\arguments{
\item{data}{A formated matrix input}

\item{path}{The target directory}

\item{name}{(Optionnal) The progeny/tree name}

\item{age}{(Optionnal) The age at which the progeny parameters were
computed from (see details)}
}
\description{
Writes the input parameter file for VPalm from a pre-formated
matrix, generally from \code{\link[=format_tree]{format_tree()}}
}
\details{
If the \code{age} argument is not provided, the function will try to extract it from
the matrix (from the \code{Modelled Months After Planting} row name).
}
