% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{coef_sample}
\alias{coef_sample}
\title{Simulation coefficient}
\usage{
coef_sample(data, epsilon)
}
\arguments{
\item{data}{The output from one of the \link[=mod_stem_diameter]{lme models}}

\item{epsilon}{epsilon value for matrix inversion}
}
\value{
The coefficients as a \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Computes the simulation coefficients from a sampling in
the variance-covariance matrix for VPalm input.
}
\details{
\code{data} must have a \code{MatG} column that represents the variance-covariance
matrix, and a \code{coef_mean} for the mean coefficient values.
Epsilon is used to avoid non semi-positive matrices, and is added to the matrix
diagonal to make it closer to the identity matrix.
The sampling is made using a normal distribution: \code{rnorm(mean= 0,sd= 1)}.
}
