% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ops.R
\name{design_plot}
\alias{design_plot}
\title{Design the plot}
\usage{
design_plot(ntrees = 2, x_dist = NULL, y_dist = NULL, x0 = 0)
}
\arguments{
\item{ntrees}{The numbers of trees required in the scene}

\item{x_dist}{The inter-row distance (m). See details.}

\item{y_dist}{The intra-row distance (m). See details.}

\item{x0}{The minimum X coordinates}
}
\value{
A list of two:
\itemize{
\item the data.frame with all information to make an Open Plant Scene file.
\item the ggplot object associated to the design
}
}
\description{
Help designing a quincunx planting pattern of palm stand.
}
\details{
If only one of \eqn{x_{dist}}{x_dist} or \eqn{y_{dist}}{y_dist} is given, the function
will compute the other distance using the following equation:
\deqn{\sqrt{z_{dist}^2-\left(\frac{z_{dist}}{2}\right)^2}}{sqrt(z_dist*z_dist-((z_dist/2)**2))}
with \eqn{z_{dist}}{z_dist} being \eqn{y_{dist}}{y_dist} if only \eqn{x_{dist}}{x_dist} is provided,
and reciprocally.
}
\note{
If \eqn{y_{dist}}{y_dist} and \eqn{x_{dist}}{x_dist} are given for the distance
between trees and not intra/inter-row distance, a transformation has to be done beforehand (see example)
}
\section{Torricity in ARCHIMED}{

If the user wants to use the ARCHIMED model for further computations, and if the torricty option
is activated, only the Voronoï sample of the design is needed (\emph{i.e.} the smaller possible
representation of the system) because the design will be virtually duplicated to infinity.
}

\examples{
# design a plot with a distance of 9.2 m between each palm trees:
  design_plot(ntrees = 5, y_dist = 9.2)

}
