% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_curvature.R
\name{deviation_A}
\alias{deviation_A}
\title{Leaf deviation}
\usage{
deviation_A(devC, Length)
}
\arguments{
\item{devC}{Deviation on each point compared to C point}

\item{Length}{The total rachis length}
}
\value{
The leaf deviation in A point
}
\description{
Compute the leaf deviation on point A
}
\examples{
\dontrun{
library(Vpalmr)
leaf_curvature(position= c(0,0.43,0.69,1),
               angC= c(27.14,27.14,27.14,27.14),
               angA= c(66.59,66.59,66.59,66.59),
               coefCurv= 3.58,
               Lenght= c(320.11,320.11,320.11,320.11))
}
}
