% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_params.R
\name{extract_params}
\alias{extract_params}
\title{Extract VPalm parameters}
\usage{
extract_params(
  data,
  model,
  leaves = 45,
  seed = sample.int(1000, 1),
  init = init_list(),
  type = c("sample", "mean")
)
}
\arguments{
\item{data}{A list of all data (generally from \code{\link[=import_data]{import_data()}})}

\item{model}{A list of the models fitted on the data (generally from
\code{\link[=mod_all]{mod_all()}})}

\item{leaves}{The number of leaves needed}

\item{seed}{The seed for random parameter generation}

\item{init}{The initialisation parameters (see details)}

\item{type}{The type of sampling performed on the parameter distribution (see
\code{\link[=coef_sample]{coef_sample()}})}
}
\value{
A list of the parameters used as VPalm input
}
\description{
Format the model outputs as a list for VPalm inputs
}
\details{
Parameters sampled include:
\itemize{
\item The rachis curvature
\item The rachis relative height
\item All lme models (see \link[=mod_stem_diameter]{models})
\item All nlme models (see \link[=mod_stem_diameter]{models})
The type parameter controls whether a sample or an average tree is needed.
}

The init parameter should be a named list of initialization
parameters as in \code{\link[=init_list]{init_list()}}. The user can use \code{\link[=init_list]{init_list()}} as a
template to change the values of the init argument (see examples).
}
\examples{
# Change the initial values of lamina angle :
ini= init_list()
ini$laminaAngle= 145
\dontrun{
 # call the function to create a VPalm input:
 extract_params(data,model, leaves= 45, seed= sample.int(1000,1),
 initialisation= ini)
}
}
