% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_raw.R
\name{extract_sheets}
\alias{extract_sheets}
\title{Extract sheets}
\usage{
extract_sheets(path, form = c("area", "development"), sheet = NULL)
}
\arguments{
\item{path}{The path to the excel file}

\item{form}{The form type (see form section)}

\item{sheet}{The sheets index on the files (see details)}
}
\value{
A data.frame with the data from all sheets (all leaves)
}
\description{
Extract the data from all forms and all sheets from the leaf area/biomass or
the development excel files
}
\details{
If the sheet is not given, the function will import data from all sheets. Otherwise,
the user can provide a given sheet, or a vector of sheets (see examples).
}
\note{
If you have an error as the following one: "JAVA_HOME cannot be determined from the Registry",
please check that you have a JAVA version that match you R one, \emph{i.e.} that your JAVA is 64-bits if
your R is 64-bits also. If JAVA is 64-bits, it should be installed in "C:/Program Files/Java",
otherwise it is located in "C:/Program Files (x86)/Java"
}
\examples{
\dontrun{
# Extract all sheets:
extract_sheets(path= "1-Data/raw/Form Archi.xlsx",form='area')

# Extract only the first sheet:
form(path= "1-Data/raw/Form Archi.xlsx", form='area', sheet= 1)

# Extract the second and third sheets:
extract_sheets(path= "1-Data/raw/Form Archi.xlsx",form='area', sheet= c(2,3))
}
}
