% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ops.R
\name{format_ops}
\alias{format_ops}
\title{Format ops information}
\usage{
format_ops(design, Progeny, map)
}
\arguments{
\item{design}{The planting design, generally computed using \code{\link[=design_plot]{design_plot()}}}

\item{Progeny}{The progeny name}

\item{map}{The tree age in month after planting}
}
\value{
A pre-formatted OPS
}
\description{
Make the planting design to the Open Plant Scene format
to prepare for OPS writing
}
