% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import all data}
\usage{
import_data(parameter, development, phylotaxy, declination, curvature,
  leaf_area, axial_angle, petiole_width, twist, map)
}
\arguments{
\item{parameter}{File path to the parameter file (ParameterSimu.csv)}

\item{development}{File path to the development file (Development_Rep4_SMSE.csv)}

\item{phylotaxy}{File path to the phylotaxy file (Stem_SMSE14.csv)}

\item{declination}{File path to the declination and leaf torsion file (A_SMSE_Nov14.csv)}

\item{curvature}{File path to the leaf curvature file (LeafCurvature_SMSE14.csv)}

\item{leaf_area}{File path to the leaf area file (LeafArea_monitoring_SMSE.csv)}

\item{axial_angle}{File path to the leaf axial angle file (LeafDispositionComp_SMSE14.csv)}

\item{petiole_width}{File path to the leaf petiole width file (Petiole_SMSE14.csv)}

\item{twist}{File path to the leaf twist file (Torsion_SMSE14.csv), used to compute
leaf nerve height.}

\item{map}{Physiological age requested in month after planting}
}
\value{
A list of the imported and pre-processed files
}
\description{
Import and pre-compute the input data for VPalm
}
\examples{
\dontrun{
Vpalmr::import_data(parameter= 'Archi/ParameterSimu.csv',
                    development= 'Archi/Development_Rep4_SMSE.csv',
                    phylotaxy= "Archi/Stem_SMSE14.csv",
                    declination= "Archi/AnglesC&A_SMSE_Nov14.csv",
                    curvature= "Archi/LeafCurvature_SMSE14.csv",
                    leaf_area= 'Archi/LeafArea_monitoring_SMSE.csv',
                    axial_angle= "Archi/LeafDispositionComp_SMSE14.csv",
                    petiole_width= "Archi/Petiole_SMSE14.csv",
                    twist= "Archi/Torsion_SMSE14.csv")
}

}
