% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{init_list}
\alias{init_list}
\title{Parameter initialization}
\usage{
init_list()
}
\arguments{
\item{seed}{The seed to use for random generation}
}
\value{
The default parameter list
}
\description{
Return the default initialisation for the parameters
}
\examples{
\dontrun{
init_list(10)
}
}
