% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scene.R
\name{make_scene}
\alias{make_scene}
\title{Make a full 3D scene}
\usage{
make_scene(data, nleaves = 45, Progeny = NULL, path, AMAPStudio,
  plant_dist = 9.2, plot_design = NULL, seed = NULL, overwrite = T,
  ntrees = NULL, progress = NULL)
}
\arguments{
\item{data}{The output from \code{\link[=compute_archi]{compute_archi()}}}

\item{nleaves}{The numbers of leaves desired for the 3D palms}

\item{Progeny}{The progeny desired. If \code{NULL} (the default), uses all progenies from \code{data}.}

\item{path}{The path for writing the scene}

\item{AMAPStudio}{The path to AMAPStudio were VPalm lives}

\item{plant_dist}{The distance between palm trees, used as y_dist in \code{\link[=design_plot]{design_plot()}}}

\item{plot_design}{The design of the plot if custom. If \code{NULL}, calls \code{\link[=design_plot]{design_plot()}}
with a quincunx disposition}

\item{seed}{The seed for random parameter generation (must be length ntrees)}

\item{overwrite}{Boolean. Should pre-existing scene files overwriten ?}

\item{ntrees}{The number of trees to be sampled for each progeny if genetic variability
is required. If \code{NULL} or \code{0}, uses only average palm trees.}

\item{progress}{A progress function (see details) Shiny progress bar if ran using Shiny application.}
}
\value{
Writes a full 3D scene with a list of OPS for each progeny and a list of OPF for each plant
of each progeny. Returns a list of three:
\itemize{
\item The OPF values for each progeny and for an average plant, as \code{\link[=extract_progeny]{extract_progeny()}} do (VPalm_parameters).
\item The data.frame of the plot design (plot_design)
\item A ggplot of the design (ggplot_design)
}
}
\description{
Sample trees for each progenies from the models provided, write parameters, call
VPalm to create OPF files, and write the OPS files to disk for each progeny.
}
\details{
As this function can take some time to run, it is possible to pass a
progress function to the \code{progress} argument as for Shiny progress bar using
\href{https://shiny.rstudio.com/articles/progress.html}{progress}. This function calls
\code{\link[=up_progress]{up_progress()}} under the hood. There are 7 calls to the progress function in total.
}
\note{
To extract only average trees from progenies, simply set \code{ntrees= 0}. To make a scene with the average
tree from all progenies, set \code{Progeny= "Average"}.
}
\examples{
\dontrun{
# Initializing:
ntrees= 0 ; nleaves= 45 ; map= 47

# Importing the inputs, and fitting the models:
Palm_Param= compute_archi(map = map, data_path = "1-Data/Archi", write_path = "3-Outputs")

# Writing the scene:
scene= make_scene(data = Palm_Param, ntrees = ntrees, nleaves = nleaves, path = "3-Outputs",
                  AMAPStudio = "../../AMAPSTUDIO/VPalm", plant_dist = 9.2)
}
}
