% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mod_all}
\alias{mod_all}
\title{Fit all models}
\usage{
mod_all(x)
}
\arguments{
\item{x}{A list of all data (generally from \code{\link[=import_data]{import_data()}})}
}
\value{
A list of all model fits
}
\description{
Fit all models
}
