% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mod_stem_diameter}
\alias{mod_stem_diameter}
\alias{mod_stem_height}
\alias{mod_rachis_length}
\alias{mod_petiole_ratio}
\alias{mod_B_position}
\alias{mod_nb_leaflet}
\alias{mod_C_declination}
\alias{mod_leaf_curvature}
\alias{mod_A_declination}
\alias{mod_A_deviation}
\alias{mod_leaflet_position}
\alias{mod_leaflet_length_B}
\alias{mod_leaflet_width_B}
\alias{mod_leaflet_length}
\alias{mod_leaflet_width}
\alias{mod_leaflet_axial_angle}
\alias{mod_leaflet_radial_angle}
\alias{mod_leaflet_type_freq}
\alias{mod_leaflet_shape}
\alias{mod_petiole_width_C}
\alias{mod_petiole_height}
\title{Fit the models}
\usage{
mod_stem_diameter(data)

mod_stem_height(data)

mod_rachis_length(data, control = nlme::lmeControl(maxIter = 5e+05, niterEM =
  25000))

mod_petiole_ratio(data)

mod_B_position(data)

mod_nb_leaflet(data)

mod_C_declination(data, control = nlme::lmeControl(maxIter = 5e+05, niterEM =
  25000))

mod_leaf_curvature(data)

mod_A_declination(data, decMaxA = 180, decSlopeA = 0.01)

mod_A_deviation(data)

mod_leaflet_position(data, Area, control = stats::nls.control(maxiter =
  5e+05))

mod_leaflet_length_B(data, control = nlme::lmeControl(maxIter = 5e+05, niterEM
  = 25000))

mod_leaflet_width_B(data, control = nlme::lmeControl(maxIter = 5e+05, niterEM
  = 25000))

mod_leaflet_length(data, control = nlme::nlmeControl(maxIter = 5e+05, niterEM
  = 25000))

mod_leaflet_width(data, control = nlme::nlmeControl(maxIter = 5e+05, niterEM =
  25000))

mod_leaflet_axial_angle(data, control = nlme::nlmeControl(maxIter = 5e+05,
  niterEM = 25000))

mod_leaflet_radial_angle(data)

mod_leaflet_type_freq(data)

mod_leaflet_shape(data)

mod_petiole_width_C(data, control = nlme::lmeControl(maxIter = 5e+05, niterEM
  = 25000))

mod_petiole_height(data, control = nlme::nlmeControl(maxIter = 5e+05, niterEM
  = 25000))
}
\arguments{
\item{data}{Input data, generally one of the objects from \code{\link[=import_data]{import_data()}}}

\item{control}{The control parameters for each kind of model}

\item{decMaxA}{The maximum allowed declination angle at the A point}

\item{decSlopeA}{The declination slope at point A}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of the model outputs, eventually grouped by Progeny
if several in input
}
\description{
These functions fit a given model from which the user can
extract parameters used as VPalm inputs
}
\section{Purpose of each function}{

\itemize{
\item \code{\link[=mod_stem_diameter]{mod_stem_diameter()}}: model the stem diameter from the rachis length using a
\code{\link[=sigmoid]{sigmoid()}}
\item \code{\link[=mod_stem_height]{mod_stem_height()}}: model the stem height using the total emitted
leaves throughout the palm life using \code{\link[=stem_height]{stem_height()}}
\item \code{\link[=mod_rachis_length]{mod_rachis_length()}}: model the leaf length from the leaf number (index of the
leaves starting from the first leaf ever emitted)
\item \code{\link[=mod_petiole_ratio]{mod_petiole_ratio()}}: model the petiole/rachis ratio by taking the average value for
the progeny at the given physiological age +/- 60. If there is no value, return the
average value by progeny for all leaves older than the reference leaf rank (>17).
\item \code{\link[=mod_nb_leaflet]{mod_nb_leaflet()}}: model the number of leaflets from the rachis length using a
\code{\link[=sigmoid]{sigmoid()}}
\item \code{\link[=mod_C_declination]{mod_C_declination()}}: model the leaf declination angle at C point from leaf rank
\item \code{\link[=mod_leaf_curvature]{mod_leaf_curvature()}}: find the leaf inclination parameters (coefCurv and decAInfl)
by minimizing the difference between a computed leaf using these parameters and the
measurements. Uses \code{\link[=optim_Leaf_Curv_Infl]{optim_Leaf_Curv_Infl()}} under the hood.
\item \code{\link[=mod_A_declination]{mod_A_declination()}}: model the leaf declination angle at A point from C angle. This
function needs the decAInfl parameter computed by \code{\link[=mod_leaf_curvature]{mod_leaf_curvature()}}, so it first
checks if the parameter is present in the input data, and computes it if absent. The user
should avoid this step by providing the parameter when possible because the
\code{\link[=mod_leaf_curvature]{mod_leaf_curvature()}} call can be computationally intensive.
\item \code{\link[=mod_A_deviation]{mod_A_deviation()}}: model the leaf deviation by using the distance from C point
and the length of the leaf between A and C point as inputs for \code{\link[=deviation_A]{deviation_A()}}, and
return the average value and standard deviation per progeny.
\item \code{\link[=mod_leaflet_position]{mod_leaflet_position()}}: model the leaflet position along the leaf rachis.
\item \code{\link[=mod_leaflet_length_B]{mod_leaflet_length_B()}}: model the leaflet length at B point
\item \code{\link[=mod_leaflet_width_B]{mod_leaflet_width_B()}}: model the leaflet width at B point
\item \code{\link[=mod_leaflet_length]{mod_leaflet_length()}}: model the leaflet relative length from its relative position
on the rachis
\item \code{\link[=mod_leaflet_width]{mod_leaflet_width()}}:  model the leaflet relative width from its relative position
on the rachis
\item \code{\link[=mod_leaflet_axial_angle]{mod_leaflet_axial_angle()}}: model the leaflet axial angle from its relative position
on the rachis
\item \code{\link[=mod_leaflet_radial_angle]{mod_leaflet_radial_angle()}}: model the leaflet radial angle from its relative position
on the rachis
\item \code{\link[=mod_leaflet_type_freq]{mod_leaflet_type_freq()}}: model the leaflet radial angle frequency along the rachis.
Leaflets can be positionned up, medium or down.
\item \code{\link[=mod_leaflet_shape]{mod_leaflet_shape()}}: model the leaflet shape
\item \code{\link[=mod_petiole_width_C]{mod_petiole_width_C()}}: model the petiole width at C point
\item \code{\link[=mod_petiole_height]{mod_petiole_height()}}: model the petiole section height using \code{\link[=section_height]{section_height()}}
}
}

\section{The models}{

\itemize{
\item \code{\link[stats:nls]{stats::nls()}} is used for \code{\link[=mod_stem_diameter]{mod_stem_diameter()}}, \code{\link[=mod_stem_height]{mod_stem_height()}},
\code{\link[=mod_nb_leaflet]{mod_nb_leaflet()}}, \code{\link[=mod_leaflet_position]{mod_leaflet_position()}}
\item \code{\link[nlme:lme]{nlme::lme()}} is used for \code{\link[=mod_rachis_length]{mod_rachis_length()}}, \code{\link[=mod_C_declination]{mod_C_declination()}},
\code{\link[=mod_leaflet_length_B]{mod_leaflet_length_B()}}, \code{\link[=mod_petiole_width_C]{mod_petiole_width_C()}}, \code{\link[=mod_petiole_height]{mod_petiole_height()}}
\item \code{\link[nlme:nlme]{nlme::nlme()}} is used for \code{\link[=mod_leaflet_length]{mod_leaflet_length()}}, \code{\link[=mod_leaflet_width]{mod_leaflet_width()}},
\code{\link[=mod_leaflet_axial_angle]{mod_leaflet_axial_angle()}}
\item \code{\link[base:mean]{base::mean()}} is used for \code{\link[=mod_petiole_ratio]{mod_petiole_ratio()}}
\item \code{\link[stats:optim]{stats::optim()}} is used by \code{\link[=mod_leaf_curvature]{mod_leaf_curvature()}}
}
}

