% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pull_lme}
\alias{pull_lme}
\alias{pull_nlme}
\title{Pull (n)lme output}
\usage{
pull_lme(data, epsilon = 10^-6, type = c("sample", "mean"))

pull_nlme(data, epsilon = 10^-6, type = c("sample", "mean"))
}
\arguments{
\item{data}{The output from one of the \link[=mod_stem_diameter]{lme models} or
\link[=mod_stem_diameter]{nlme models}}

\item{epsilon}{epsilon value for matrix inversion}

\item{type}{The type of sampling performed on the parameter distribution (see
\code{\link[=coef_sample]{coef_sample()}})}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with model outputs
}
\description{
Format lme or nlme outputs in a standard \code{\link[tibble:tibble]{tibble::tibble()}} and computes
some usefull informations such as covariance matrices. Carefull, this function is
poorly designed for the moment and is only applicable on the (n)lme used in the context
of the package because only two or three factors (lme and nlme resp.) can be used.
}
\details{
Epsilon is used to avoid non semi-positive matrices, and is added to the matrix
diagonal to make it closer to the identity matrix
}
