% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_dimensions.R
\name{shape_beta}
\alias{shape_beta}
\title{Beta function for leaflet shape}
\usage{
shape_beta(x, xm, ym)
}
\arguments{
\item{x}{Relative position on the leaflet}

\item{xm}{Position of the maximum leaflet width (0-1)}

\item{ym}{Shape factor}
}
\value{
The relative width of the leaflet at x position
}
\description{
The function computes the relative width of the leaflet at all x
positions. The outputed width is relative to the leaflet max width.
}
