% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_inputs.R
\name{test_Area}
\alias{test_Area}
\title{Test the Area data.frame}
\usage{
test_Area(Area)
}
\arguments{
\item{Area}{The Area data.frame}
}
\value{
\code{TRUE} if the file had no major problem, or a special output that points to the potential
error
}
\description{
Performs several tests on the Area file to check for potential errors or typos
}
\examples{
\dontrun{
library(Vpalmr)
Area= data.table::fread('LeafArea_monitoring_SMSE.csv',data.table = F,dec=',')
test_Area(Area= Area)
}
}
