% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPalm_list.R
\name{VPalm_list}
\alias{VPalm_list}
\title{VPalm input}
\usage{
VPalm_list(data, model, nb_leaves = 45, seed = sample.int(1000, 1))
}
\arguments{
\item{data}{A list of all data (generally from \code{\link[=import_data]{import_data()}})}

\item{model}{A list of the models fitted on the data (generally from
\code{\link[=mod_all]{mod_all()}})}

\item{nb_leaves}{The number of leaves needed}

\item{seed}{The seed for random parameter generation}
}
\value{
A list of the parameters used as VPalm input
}
\description{
Format the model outputs as a list for VPalm inputs
}
