% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_VPalm.R
\name{write_progeny}
\alias{write_progeny}
\title{Writes all progenies VPalm inputs}
\usage{
write_progeny(data, path)
}
\arguments{
\item{data}{A pre-formated \code{\link[=format_progeny]{format_progeny()}} output list of matrices}

\item{path}{The path of the directory to write in}
}
\description{
Applies \code{\link[=write_tree]{write_tree()}} successively to a \code{\link[=format_progeny]{format_progeny()}} output to
write all trees from all progenies to a VPalm input file.
}
