% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_VPalm.R
\name{write_progeny}
\alias{write_progeny}
\title{Writes all progenies VPalm inputs}
\usage{
write_progeny(data, path = getwd(), wforce = FALSE, overwrite = TRUE,
  verbose = TRUE)
}
\arguments{
\item{data}{A pre-formated \code{\link[=format_progeny]{format_progeny()}} output list of matrices}

\item{path}{The path of the directory to write in}

\item{wforce}{Boolean. If \code{TRUE} (the default), continue to write files even if one
returned an error, and return warnings instead. If \code{FALSE}, stop at first error.}

\item{overwrite}{Boolean. Overwrite a pre-existing VPalm input file if set to \code{TRUE}}

\item{verbose}{Prints a messages during writing. Especially usefull for debugging.}
}
\value{
A matrix informing about the success of the writing of the files. All values
should be \code{TRUE} if all file were successfully written. The text files written are named
as follows: "{Progeny}_{Tree}\emph{MAP}{X}.txt", where Progeny is the progeny name, Tree is the tree
name ("Average" if not sampled), and X the number of months after planting.
If average data is present in \code{data}, the function writes it along the others, and name the output
"Average_Average_MAP_X", with X being the month after planting.
}
\description{
Applies \code{\link[=write_tree]{write_tree()}} successively to a \code{\link[=format_progeny]{format_progeny()}} output to
write all trees from all progenies to a VPalm input file.
}
