% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_VPalm.R
\name{write_tree}
\alias{write_tree}
\title{Write VPalm tree parameter file}
\usage{
write_tree(
  data,
  path = getwd(),
  name = "vpalm_input",
  age = NULL,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A formated matrix input}

\item{path}{The target directory}

\item{name}{(Optionnal) The progeny/tree name}

\item{age}{(Optionnal) The age at which the progeny parameters were
computed from (see details)}

\item{overwrite}{Boolean. Overwrite a pre-existing VPalm input file if set to \code{TRUE}}

\item{verbose}{Prints a message if the writing is successfull}
}
\value{
\code{TRUE} if the file was successfully written, \code{FALSE} if any error
was encountered during the process.
}
\description{
Writes the input parameter file for VPalm from a pre-formated
matrix, generally from \code{\link[=format_tree]{format_tree()}}
}
\details{
If the \code{age} argument is not provided, the function will try to extract it from
the matrix (from the \verb{Modelled Months After Planting} row name).
}
